+++ nflow-engine/src/test/java/io/nflow/engine/internal/executor/WorkflowDispatcherTest.java
@@ -81,7 +81,6 @@ public class WorkflowDispatcherTest {
     env.setProperty("nflow.executor.stateProcessingRetryDelay.seconds", "1");
     env.setProperty("nflow.executor.stateSaveRetryDelay.seconds", "60");
     env.setProperty("nflow.executor.stateVariableValueTooLongRetryDelay.minutes", "60");
-    env.setProperty("nflow.executor.fetchChildWorkflowIds", "false");
     env.setProperty("nflow.db.workflowInstanceType.cacheSize", "10000");
     when(executorDao.isTransactionSupportEnabled()).thenReturn(true);
     when(executorDao.isAutoCommitEnabled()).thenReturn(true);